Switch from int to long (#8614)
Interlocked has poor codegen for non-word size, so fix that.

+++ main/src/core/MonoDevelop.Core/MonoDevelop.Projects.MSBuild/RemoteBuildEngine.cs
@@ -42,7 +42,7 @@ namespace MonoDevelop.Projects.MSBuild
 	{
 		RemoteProcessConnection connection;
 		bool alive = true;
-		static int count;
+		static long count;
 		int busy;
 		int buildSessionLoggerId;
 		CancellationTokenSource disposalCancelSource;
